requirements.txt
streamlit
transformers
torch
pdfminer.six
sentence-transformers
scikit-learn
